+++ plugins/module_utils/infinibox.py
@@ -7,7 +7,10 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 from ansible.module_utils.six import raise_from
-import ansible.module_utils.errors
+try:
+    import ansible.module_utils.errors
+except ImportError:
+    import errors
 
 try:
     from infinisdk import InfiniBox, core